os-592 Implemented mechanism for custom checks
This commit adds the `extra` parameter to
sys11sensu::profile::server::checks::extra. This parameter takes a hash
keyed by check names with the values being parameters to the
sensu::check resource. You can use this parameter to specify additional
checks to be run on a Sensu server (e.g. a check for monitoring a remote
web server's reachability.